pyproject.toml
[project]
name = "swm-x-gemini-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"google-genai",
	"google-api-python-client",
	"google-auth-oauthlib",
	"markdown",
	"python-dotenv",
]
